Handle server push via grunt
Setup grunt to handle pushing to server, as well as local dist and staging.

Grunt Tasks Overview
--------------------

grunt dev: Push to ditribution folder html/dist/ and to
/var/www/fiveormore-xxxxx

grunt prod: push to server, into fiveormore.xxxx.

Note that prod task does not remove the existing symlink on server. It provides
an ssh line to copy and paste to delete old symlink and link to new directory.
This is the point that the new server goes live. Before that it can be tested
at server-url/fiveormore-xxxx/. (Link provided at end of grunt tasks).